# src/controllers/songController.js
const { StatusCodes } = require("http-status-codes");
const User = require('../models/User')
const Album = require("../models/Album");
const Playlist = require("../models/Playlist");
const Artist = require("../models/Artist");
const Song = require("../models/Song")
const { uploadToCloudinary } = require("../utils/cloudinaryUpload");
const mm = require('music-metadata');

const createSong = async (req, res) => {
    try {
        const { title, artistId, albumId, genre, lyrics, isExplicit, featuredArtists } = req.body
        const artist = await Artist.findById(artistId)
        if (!artist) {
            res.status(StatusCodes.NOT_FOUND);
            throw new Error("Artist not found");
        }
        if (albumId) {
            const album = await Album.findById(albumId)
            if (!album) {
                res.status(StatusCodes.NOT_FOUND);
                throw new Error("Album not found");
            }
        }
        if (!req.files || !req.files.audio) {
            res.status(StatusCodes.BAD_REQUEST);
            throw new Error("Audio file is required");
        }
        const audioPath = req.files.audio[0].path
        const metadata = await mm.parseFile(audioPath)
        const durationInSeconds = Math.floor(metadata.format.duration)

        const audioResult = await uploadToCloudinary(audioPath, "spotify/songs")

        let coverImageUrl = ""
        if (req.files && req.files.cover) {
            const imageResult = await uploadToCloudinary(req.files.cover[0].path, "spotify/covers")
            coverImageUrl = imageResult.secure_url
        }

        // let parsedDuration = duration;
        // if (typeof duration === "string" && duration.includes(":")) {
        //     const [minutes, seconds] = duration.split(":").map(Number);
        //     parsedDuration = minutes * 60 + seconds;
        // }

        const song = await Song.create({
            title,
            artist: artistId,
            album: albumId || null,
            duration: durationInSeconds,
            audioUrl: audioResult.secure_url,
            genre,
            lyrics,
            isExplicit: isExplicit === 'true',
            // featuredArtists: featuredArtists ? JSON.parse(featuredArtists) : [],
            featuredArtists: featuredArtists || [],
            coverImage: coverImageUrl
        })

        artist.songs.push(song._id)
        await artist.save()

        if (albumId) {
            const album = await Album.findById(albumId)
            album.songs.push(song._id)
            await album.save()
        }
        res.status(StatusCodes.CREATED).json(song)
    } catch (error) {
        console.error("Error in createSong", error)
        res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ message: "Something went wrong", error: error.message })
    }
}

const getSongs = async (req, res) => {
    try {
        const { genre, artist, search, page = 1, limit = 10 } = req.query

        const filter = {}
        if (genre) {
            filter.genre = genre
        }
        if (artist) {
            filter.artist = artist
        }
        if (search) {
            filter.$or = [
                {
                    title: {
                        $regex: search,
                        $options: "i"
                    }
                },
                {
                    genre: {
                        $regex: search,
                        $options: "i"
                    }
                }
            ]
        }
        const count = await Song.countDocuments(filter)
        const skip = (parseInt(page) - 1) * parseInt(limit)
        const songs = await Song.find(filter)
            .sort({ releaseDate: -1 })
            .limit(parseInt(limit))
            .skip(skip)
            .populate("artist", "name image")
            .populate("album", "name coverImage")
            .populate("featuredArtists", "name")

        res.status(StatusCodes.OK).json({
            songs,
            page: parseInt(page),
            pages: Math.ceil(count / parseInt(limit)),
            totalSongs: count
        })
    } catch (error) {
        console.error("Error in getSongs", error)
        res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ message: "Something went wrong", error: error.message })
    }
}

const getSongById = async (req, res) => {
    try {
        const id = req.params.id
        const existingSong = await Song.findById(id)
            .populate("artist", "name image bio")
            .populate("album", "title coverImage releasedDate")
            .populate("featuredArtists", "name image")

        if (!existingSong) {
            res.status(StatusCodes.NOT_FOUND)
            throw new Error("Song not found");
        }
        existingSong.plays += 1
        await existingSong.save()
        res.status(StatusCodes.OK).json({ existingSong })
    } catch (err) {
        console.error("Error in getSongsById", error)
        res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ message: "Something went wrong", error: error.message })
    }
}

const updateSong = async (req, res) => {
    try {
        const { title, artistId, albumId, duration, genre, lyrics, isExplicit, featuredArtists } = req.body
        const song = await Song.findById(req.params.id)
        if (!song) {
            res.status(StatusCodes.NOT_FOUND)
            throw new Error("Song not found")
        }
        song.title = title || song.title
        song.album = albumId || song.album
        song.genre = genre || song.genre
        song.lyrics = lyrics || song.lyrics
        song.artist = artistId || song.artist
        song.duration = duration || song.duration
        // song.isExplicit = isExplicit !== undefined ? isExplicit = "true" : song.isExplicit
        if (isExplicit !== undefined) {
            song.isExplicit = (isExplicit === 'true' || isExplicit === true);
        }
        song.featuredArtists = featuredArtists ? JSON.parse(featuredArtists) : song.featuredArtists
        if (req.files && req.files.cover) {
            const imageResult = await uploadToCloudinary(
                req.files.cover[0].path,
                "spotify/covers"
            );
            song.coverImage = imageResult.secure_url;
        }
        if (req.files && req.files.audio) {
            const audioResult = await uploadToCloudinary(
                req.files.audio[0].path,
                "spotify/songs"
            );
            song.audioUrl = audioResult.secure_url;
        }

        const updatedSong = await song.save();
        res.status(StatusCodes.OK).json(updatedSong);
    } catch (error) {
        console.error("Error in updateSong", error)
        res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ message: "Something went wrong", error: error.message })
    }
}

const deleteSong = async (req, res) => {
    try {
        const song = await Song.findById(req.params.id)
        if (!song) {
            res.status(StatusCodes.NOT_FOUND)
            throw new Error("Song not found")
        }

        await Artist.updateOne({ _id: song.artist }, { $pull: { songs: song._id } })

        if (song.album) {
            await Album.updateOne({ _id: song.album }, { $pull: { songs: song._id } })
        }

        await song.deleteOne()
        res.status(StatusCodes.OK).json({ message: "Song deleted" });
    } catch (error) {
        console.error("Error in deleteSong", error)
        res.status(StatusCodes.INTERNAL_SERVER_ERROR).json({ message: "Something went wrong", error: error.message })
    }
}

module.exports = { createSong, getSongs, getSongById, updateSong, deleteSong }