# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My User API",
    "version": "1.0.0",
    "description": "API for user management, authentication, etc."
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token in the format: Bearer <token>"
      }
    }
  },
  "paths": {
...
    "/api/songs": {
      "post": {
        "tags": [
          "Songs"
        ],
        "summary": "Create a new song - Admin only",
        "description": "Uploads a new song with audio, optional cover image, and metadata.",
        "operationId": "createSong",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "example": "Shape of You"
                  },
                  "artistId": {
                    "type": "string",
                    "example": "64f8a2b3c1e5f8a2b3c1e5f9"
                  },
                  "albumId": {
                    "type": "string",
                    "example": "64f8b3a1d2e5a8b3c1e5f8d2",
                    "nullable": true
                  },
                  "genre": {
                    "type": "string",
                    "example": "Pop"
                  },
                  "lyrics": {
                    "type": "string",
                    "example": "The club isn't the best place to find a lover..."
                  },
                  "isExplicit": {
                    "type": "boolean",
                    "example": false
                  },
                  "featuredArtists": {
                    "type": "string",
                    "description": "JSON array of featured artist IDs",
                    "example": "[\"64f8c3a1d2e5a8b3c1e5f8a9\"]"
                  },
                  "audio": {
                    "type": "string",
                    "format": "binary",
                    "description": "Audio file (required, e.g. mp3, wav, m4a)"
                  },
                  "cover": {
                    "type": "string",
                    "format": "binary",
                    "description": "Cover image (optional)"
                  }
                },
                "required": [
                  "title",
                  "artistId",
                  "audio"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Song created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Song"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing audio file or invalid input"
          },
          "404": {
            "description": "Artist or album not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "get": {
        "tags": [
          "Songs"
        ],
        "summary": "Get all songs",
        "description": "Retrieve a list of songs with optional filtering by genre, artist, search, and pagination.",
        "operationId": "getSongs",
        "parameters": [
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter songs by genre"
          },
          {
            "name": "artist",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter songs by artist ID"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Search songs by title or genre (case-insensitive)"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            },
            "description": "Page number for pagination"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 10
            },
            "description": "Number of songs per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of songs retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "songs": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Song"
                      }
                    },
                    "page": {
                      "type": "integer",
                      "example": 1
                    },
                    "pages": {
                      "type": "integer",
                      "example": 5
                    },
                    "totalSongs": {
                      "type": "integer",
                      "example": 42
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/artists": {
      "post": {
        "tags": [
          "Artists"
        ],
        "summary": "Create a new artist",
        "description": "Only admins can create artists. Upload artist details along with an optional image.",
        "operationId": "createArtist",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Taylor Swift"
                  },
                  "bio": {
                    "type": "string",
                    "example": "American singer-songwriter. Known for narrative songwriting."
                  },
                  "genres": {
                    "type": "string",
                    "example": "Pop, Country"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary",
                    "description": "Artist image file"
                  }
                },
                "required": [
                  "name",
                  "bio",
                  "genres"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artist created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artist"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - missing fields or artist already exists"
          },
          "401": {
            "description": "Unauthorized - no token provided"
          },
          "403": {
            "description": "Forbidden - user is not an admin"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  }
}